Assignment 2/requirements.txt
lightning==1.4.0
wandb
torchaudio
pandas
numpy
matplotlib
seaborn
scikit-learn
pytorch==1.8.1